+++ projects/safe/src/lib/services/form.service.ts
@@ -2,7 +2,7 @@ import { Inject, Injectable } from '@angular/core';
 import * as SurveyKo from 'survey-knockout';
 import * as Survey from 'survey-angular';
 import { initCreatorSettings } from '../survey/creator';
-import { initCustomWidgets } from '../survey/init';
+import { initCustomSurvey } from '../survey/init';
 import { DomService } from './dom.service';
 import { MatDialog } from '@angular/material/dialog';
 import { Apollo } from 'apollo-angular';
@@ -39,7 +39,7 @@ export class SafeFormService {
     public authService: SafeAuthService
   ) {
     // === CUSTOM WIDGETS / COMPONENTS ===
-    initCustomWidgets(
+    initCustomSurvey(
       SurveyKo,
       domService,
       dialog,
@@ -51,7 +51,7 @@ export class SafeFormService {
     // === CREATOR SETTINGS ===
     initCreatorSettings(SurveyKo);
     // === CUSTOM WIDGETS / COMPONENTS ===
-    initCustomWidgets(
+    initCustomSurvey(
       Survey,
       domService,
       dialog,
